Develop an illustration of MBD.
This commit adds a directory `model` in which we illustrate model-based
design (MBD).

The models are briefly described in the `README.md` in `model`. They
include:
* an NXT motor simulation model
* a simple open-loop harness for the NXT motor model
* a steering controller software-design model
* a closed-loop model incorporating the NXT motor, the steering
  controller, and simulation inputs

The NXT motor and steering controller models are both ready for code
generation using QGen. They are configured using `init.m`, which
introduces types and parameters bound to the Ada code that permit
interfacing with the rest of the RC Car Demo.

(Note: this is currently imperfect due to the way in which the
`Steering_Offset` is handled.)

This commit also introduces an Ada simulation harness that executes the
NXT motor and steering controller generated code and the original Ada
steering controller. This is described in the `README.md` in the
`model/harness` directory.