Improve MBD paradigm with model args and project.
This commit introduces a Simulink project to organize the Simulink
models and automate their initialization.

The models have been updated to take advantage of the Model Workspace:
this keeps the MATLAB Workspace clean and improves the organization of
code generated by QGen.

To allow more elegant / better-encapsulated configuration of the
`Steering_Offset` parameter, it has been converted to a Model Argument
for both the `Steering_Controller` and `NXT_Motor` models. This allows
callers of the generated code to set this value as a parameter in the
`initStates` procedure, which is far more elegant than using a shared
constant (or function).

The README has been updated to describe how to use the models with the
Simulink project.